NextCompanyBOTzip/NextCompanyBOTzip/requirements.txt
discord.py
python-dotenv
aiohttp
beautifulsoup4
duckduckgo-search
google-generativeai
aiohttp
beautifulsoup4
discord.py
duckduckgo-search
google-generativeai
python-dotenv